Compelte folder refactor to match method type
