+++ carbonmark-api/src/routes/categories/get.schema.ts
